Refactor org.enso.projectmanager.boot.configuration
Incremenetal compiler has occasional issues with nested classes.
Refactored classes to their own package so as to avoid spurious errors.
