+++ lib/scala/project-manager/src/main/java/org/enso/projectmanager/infrastructure/migration/ProjectsMigration.java
@@ -9,7 +9,7 @@ import java.util.HashSet;
 import java.util.Set;
 import org.apache.commons.io.FileUtils;
 import org.enso.desktopenvironment.Platform;
-import org.enso.projectmanager.boot.configuration;
+import org.enso.projectmanager.boot.configuration.StorageConfig;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
@@ -19,7 +19,7 @@ public final class ProjectsMigration {
 
   private ProjectsMigration() {}
 
-  public static void migrate(configuration.StorageConfig storageConfig) {
+  public static void migrate(StorageConfig storageConfig) {
     var oldProjectsPath =
         Platform.getOperatingSystem()
             .getDirectories()
