+++ lib/scala/project-manager/src/main/scala/org/enso/projectmanager/boot/configuration/MainProcessConfig.scala
@@ -1,0 +1,23 @@
+package org.enso.projectmanager.boot.configuration
+
+import org.slf4j.event.Level
+
+import java.nio.file.Path
+import scala.concurrent.duration.FiniteDuration
+
+/** The options supplied (e.g. with the command line options when starting the
+  * main project manager process.
+  *
+  * @param logLevel the logging level
+  * @param profilingPath the path to the profiling out file
+  * @param profilingTime the time limiting the profiling duration
+  * @param jvmMode if true, enables JVM mode
+  * @param extraEnv extra environment variables
+  */
+case class MainProcessConfig(
+  logLevel: Level,
+  profilingPath: Option[Path],
+  profilingTime: Option[FiniteDuration],
+  jvmMode: Boolean,
+  extraEnv: Seq[(String, String)]
+)
