+++ lib/scala/project-manager/src/main/scala/org/enso/projectmanager/boot/configuration/MetadataStorageConfig.scala
@@ -1,0 +1,11 @@
+package org.enso.projectmanager.boot.configuration
+
+/** A configuration object for metadata storage.
+  *
+  * @param projectMetadataDirectory a directory name containing project metadata
+  * @param projectMetadataFileName a name of project metadata file
+  */
+case class MetadataStorageConfig(
+  projectMetadataDirectory: String,
+  projectMetadataFileName: String
+)
